:recycle: Update extracted showWelcomeScreen into init.consts
